Closes #8943 COMPAT: periods needs coercion to np.int64
In _generate_regular_range, we explicitly cast `periods` to np.int64 to prevent
overflow in the function tseries.index._generate_regular_range.
Note: we don't bother casting it for the function call generate_range since
there's no danger of overflow in that function.

+++ pandas/tseries/index.py
@@ -1781,11 +1781,11 @@ def _generate_regular_range(start, end, periods, offset):
             tz = start.tz
         elif start is not None:
             b = Timestamp(start).value
-            e = b + periods * stride
+            e = b + np.int64(periods) * stride
             tz = start.tz
         elif end is not None:
             e = Timestamp(end).value + stride
-            b = e - periods * stride
+            b = e - np.int64(periods) * stride
             tz = end.tz
         else:
             raise NotImplementedError
